fix rules instantiation in Loader

+++ mock-esp/src/OutputHandler.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include "OutputHandler.h"
 
-OutputHandler::OutputHandler(OutInRule rule) { 
+OutputHandler::OutputHandler(OutInRule rule) {
   this->rule = rule;
   this->count = 0;
   this->startTime = 0;
@@ -10,13 +10,13 @@ OutputHandler::OutputHandler(OutInRule rule) {
 }
 
 void OutputHandler::update() {
-  if(isDone()) return;
+  if (isDone()) return;
   if (this->startTime == 0) this->startTime = millis();
-  if (!this->hasStarted && this->rule.getTimeout() < millis() - this->startTime) {
+  if (!this->hasStarted && this->rule.getTimeout() < (long)(millis() - this->startTime)) {
     sendRequest();
     this->hasStarted = true;
   }
-  if (hasStarted && this->rule.getInterval() < millis() - this->lastTime) {
+  if (hasStarted && this->rule.getInterval() < (long)(millis() - this->lastTime)) {
     sendRequest();
   }
 }
